Time: 521 ms (9.16%) | Memory: 18.5 MB (50.47%) - LeetSync

+++ 240-search-a-2d-matrix-ii/search-a-2d-matrix-ii.cpp
@@ -3,15 +3,15 @@ class Solution
     public:
         bool searchMatrix(vector<vector < int>> &matrix, int target)
         {
-            for (int i = 0; i < matrix.size(); ++i)
+            for (short i = 0; i < matrix.size(); ++i)
             {
                 if (matrix[i][matrix[i].size() / 2] == target) return true;
 
-                for (int j = 0; j < matrix[i].size(); ++j)
-                    {
-                        if (matrix[i][j] == target) return true;
-                        else if (matrix[i][j] > target) break;
-                    }
+                for (short j = 0; j < matrix[i].size(); ++j)
+                {
+                    if (matrix[i][j] == target) return true;
+                    else if (matrix[i][j] > target) break;
+                }
             }
             return false;
         }
